Pantalla de Usuario y de Login combinadas. Colores Unificados. No se permite rotar ninguna de las pantallas.
