+++ app/src/main/AndroidManifest.xml
@@ -10,20 +10,18 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity android:name=".Login"
-            android:label="@string/app_name" >
+            android:label="@string/app_name"
+            android:screenOrientation="portrait">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-<<<<<<< HEAD
-        <activity android:name=".Usuario"></activity>
-=======
         <activity
-            android:name=".Dummy"
-            android:label="@string/app_name" />
->>>>>>> master
+            android:name=".Usuario"
+            android:label="@string/app_name"
+            android:screenOrientation="portrait"/>
     </application>
