+++ app/src/main/java/com/policonnect/team/policonnect20/Usuario.java
@@ -2,10 +2,10 @@ package com.policonnect.team.policonnect20;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
-import android.widget.TextView;
+import android.widget.Toast;
 
 public class Usuario extends AppCompatActivity {
-
+    private static long back_pressed;
     //private TextView nombreUsuario;
 
     @Override
@@ -14,5 +14,19 @@ public class Usuario extends AppCompatActivity {
         setContentView(R.layout.activity_usuario);
         //nombreUsuario = (TextView)findViewById(R.id.nombreUsuario);
         //nombreUsuario.setText("Alejandro Arevalo");
+    }
+
+    /**
+     * Este método asegura que el usuario salga de la aplicación cuando este oprima e botón "atrás" dos
+     * veces con un lapso de tiempo de menos de 2 segundos.
+     * Sobreescribe en metodo onBackPressed
+     **/
+    @Override
+    public void onBackPressed() {
+        if (back_pressed + 2000 > System.currentTimeMillis()) {
+            System.exit(0);
+        } else
+            Toast.makeText(getBaseContext(), R.string.toast_backagain, Toast.LENGTH_SHORT).show();
+        back_pressed = System.currentTimeMillis();
     }
 }
