+++ app/src/main/res/values/styles.xml
@@ -3,16 +3,11 @@
     <!-- Base application theme. -->
     <style name="AppTheme" parent="Theme.AppCompat.Light.NoActionBar">
         <!-- Customize your theme here. -->
-<<<<<<< HEAD
-        <item name="colorPrimary">@color/colorPrimary</item>
+        <item name="colorPrimary">@color/blue2</item>
         <item name="colorPrimaryDark">@color/colorHeader</item>
-        <item name="colorAccent">@color/colorAccent</item>
-=======
-        <item name="colorPrimary">@color/blue1</item>
-        <item name="colorPrimaryDark">@color/blue2</item>
         <item name="colorAccent">@color/blueAccent</item>
 
->>>>>>> master
+
     </style>
 
 </resources>
